requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
bcrypt=4.0.1=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.08.22=haa95532_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
dataclasses-json=0.6.2=pypi_0
distro=1.8.0=pypi_0
fastapi=0.104.1=pypi_0
frozenlist=1.4.0=pypi_0
greenlet=3.0.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.25.1=pypi_0
idna=3.4=pypi_0
jinja2=3.1.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
langchain=0.0.335=pypi_0
langsmith=0.0.64=pypi_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
numpy=1.26.2=pypi_0
openai=1.2.4=pypi_0
openssl=3.0.12=h2bbff1b_0
packaging=23.2=pypi_0
passlib=1.7.4=pypi_0
pip=23.3=py311haa95532_0
pydantic=2.5.0=pypi_0
pydantic-core=2.14.1=pypi_0
python=3.11.5=he1021f5_0
python-dotenv=1.0.0=pypi_0
python-multipart=0.0.6=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py311haa95532_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.23=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.27.0=pypi_0
tenacity=8.2.3=pypi_0
tk=8.6.12=h2bbff1b_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
uvicorn=0.24.0.post1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py311haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
